+++ geoimagenet_api/__about__.py
@@ -1,3 +1,3 @@
-__version__ = "0.5.0"
+__version__ = "0.7.0"
 __author__ = "David Caron"
 __email__ = "david.caron@crim.ca"