+++ geoimagenet_api/openapi_schemas.py
@@ -76,7 +76,7 @@ class ValidationPost:
 
 
 @dataclass
-class AnnotationCount:
+class AnnotationCountPerStatus:
     new: int = field(default=0)
     pre_released: int = field(default=0)
     released: int = field(default=0)
@@ -86,7 +86,7 @@ class AnnotationCount:
     deleted: int = field(default=0)
 
     def __add__(self, other):
-        return AnnotationCount(
+        return AnnotationCountPerStatus(
             new=self.new + other.new,
             pre_released=self.pre_released + other.pre_released,
             released=self.released + other.released,
@@ -108,17 +108,18 @@ class AnnotationCount:
 
 
 @dataclass
-class AnnotationCounts:
-    taxonomy_class_id: int
-    counts: AnnotationCount
-
-
-@dataclass
 class AnnotationProperties:
     annotator_id: int
     taxonomy_class_id: int
     image_name: str
     status: str = field(default=AnnotationStatus.new.name)
+
+
+@dataclass
+class AnnotationStatusUpdate:
+    annotation_ids: List[str] = field(default=None)
+    taxonomy_class_id: int = field(default=None)
+    with_taxonomy_children: bool = field(default=True)
 
 
 @dataclass